refactored pkg names
