+++ cmd/manager/main.go
@@ -22,8 +22,8 @@ import (
 	"fmt"
 	"github.com/openebs/node-disk-manager/pkg/setup"
 	"github.com/openebs/node-disk-manager/pkg/upgrade"
-	"github.com/openebs/node-disk-manager/pkg/upgrade/v040-041"
-	"github.com/openebs/node-disk-manager/pkg/upgrade/v041-042"
+	"github.com/openebs/node-disk-manager/pkg/upgrade/v040_041"
+	"github.com/openebs/node-disk-manager/pkg/upgrade/v041_042"
 	"os"
 	"runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
